requirements.txt
dwave-ocean-sdk>=6.0
numpy
dwave-neal
